pyproject.toml
[project]
name = "caption_via_translation"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate==1.3.0",
    "clip",
    "ctranslate2>=4.6.3",
    "einops>=0.8.1",
    "evaluate>=0.4.6",
    "faiss-cpu>=1.13.0",
    "fvcore>=0.1.5.post20221221",
    "matplotlib>=3.10.7",
    "multilingual-clip>=1.0.10",
    "nltk>=3.9.2",
    "numpy>=1.26.4",
    "open-clip-torch>=3.2.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pycocoevalcap>=1.2",
    "sacrebleu>=2.5.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sglang>=0.5.2",
    "stanza>=1.11.0",
    "statsmodels>=0.14.5",
    "tabulate>=0.9.0",
    "torch>=2.9.0",
    "tqdm>=4.67.1",
    "transformers==4.48.1",
    "unbabel-comet==2.2.7",
    "wandb>=0.22.3",
    "webdataset>=1.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
